[llvm][tmp] Make deserialization listener multiplexer public
